Improved function to determine whether line intersects rectangle.
Added functionality to check for any arbitrary line by discretising into 5 straight lines and comparing these.

Removed "rectParentNode" element - rects are now appended to the element that .call is used for. In turn, added "rectRelativeToNode" to allow for use when transforms are applied to all elements (commonly pie charts and geo charts).