Added checks to ensure required settings are supplied.
Added a check so that, when selecting inside closed paths, the path itself is also selected without the need for the rectangle to intersect any of the path's borders.

Scanning now matches transforms applied to the rect and the things under the rect - effectively this makes drag selecting now work when, for example, zooming in to a map.